fire-risk-app/requirements-app.txt
folium
flask
pandas
geopandas
scikit-learn
xgboost
google-cloud-storage
